requirements.txt
Django==5.0.4
python-telegram-bot==20.3
asgiref==3.8.1